backend/requirements-production.txt
# Production requirements for AI Financial Planning Backend
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Data validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client & CORS
httpx==0.25.2
python-dateutil==2.8.2

# Environment & Configuration
python-dotenv==1.0.0

# Optional for enhanced features
redis==5.0.1
celery==5.3.4

# Development tools (optional)
pytest==7.4.3
pytest-asyncio==0.21.1